# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "PixEz Sync 后端同步服务 API。除 /mirror/** Pixiv 形态镜像接口外，系统 API 均返回 {success,message,data} 标准响应。",
        "title": "PixEz Sync Server API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/pixez/users/{pixiv_user_id}/sync-data": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Fetches all or selected synced data tables for a Pixiv user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sync Data"
                ],
                "summary": "Get synced user data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pixiv user ID",
                        "name": "pixiv_user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated table names, for example ban_comments,illust_histories",
                        "name": "tables",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserDataPayload"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Replaces the provided synced data tables for a Pixiv user in a single database transaction.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sync Data"
                ],
                "summary": "Replace synced user data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pixiv user ID",
                        "name": "pixiv_user_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Synced user data payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserDataPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/pixez/users/{pixiv_user_id}/sync-data/hashes": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "description": "Returns MD5 checksums for each synced table of a Pixiv user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Sync Data"
                ],
                "summary": "Get synced data hashes",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pixiv user ID",
                        "name": "pixiv_user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.UserDataHashes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "model.BanComment": {
            "type": "object",
            "properties": {
                "comment_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.BanIllust": {
            "type": "object",
            "properties": {
                "illust_id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.BanTag": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "translate_name": {
                    "type": "string"
                }
            }
        },
        "model.BanUser": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
...
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "model.IllustHistory": {
            "type": "object",
            "properties": {
                "illust_id": {
                    "type": "integer"
                },
                "picture_url": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "model.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "model.NovelHistory": {
            "type": "object",
            "properties": {
                "novel_id": {
                    "type": "integer"
                },
                "picture_url": {
                    "type": "string"
                },
                "time": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "model.TagHistory": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "translated_name": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "model.UserDataHashes": {
            "type": "object",
            "additionalProperties": {
                "type": "string"
            }
        },
        "model.UserDataPayload": {
            "type": "object",
            "properties": {
                "ban_comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BanComment"
                    }
                },
                "ban_illusts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BanIllust"
                    }
                },
                "ban_tags": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BanTag"
                    }
                },
                "ban_users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.BanUser"
                    }
                },
                "illust_histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.IllustHistory"
                    }
                },
                "novel_histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.NovelHistory"
                    }
                },
                "tag_histories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TagHistory"
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}